requirements.txt
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-vision==0.41.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonify==0.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pyrsistent==0.15.7
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
